rename cryptoDonation to OneTimeDonation
